added new model of the entities

+++ src/model/Employee.java
@@ -5,7 +5,7 @@ import modelutil.Person;
 public class Employee extends Person{
 
 	private final int employeeID;
-	private final int salary;
+	private final float salary;
 
 	public Employee(final String firstName, final String secondName, final String ficalCode,
 			final int employeeID, final int salary) {
@@ -20,7 +20,7 @@ public class Employee extends Person{
 		return employeeID;
 	}
 
-	public int getSalary() {
+	public float getSalary() {
 		return salary;
 	}
